# cmd/filmlibrary/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Film Library Application",
        "title": "Film Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/actors/{id}": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Delete an actor by path parameter 'id'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Delete an actor by path parameter 'id'",
                "operationId": "delete-actor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Actors`s id that needs to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.deleteActorResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Update fully or partially an actor by path parameter 'id'",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Update fully or partially an actor by path parameter 'id'",
                "operationId": "update-actor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Actors`s id that needs to be updated",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Actor object with values that will be updated",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.updateActorRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.updateActorResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            }
        },
...
        "/films/searches": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Search films by film title or/and actor name with optional 'limit' and 'offset' query parameters.\nIf 'film-title' and 'actor-name' are empty, than non-empty list of films with max length = 'limit' will be returned",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "films"
                ],
                "summary": "Search films by film`s title or/and actor`s name with optional 'limit' and 'offset' query parameters",
                "operationId": "search-films",
                "parameters": [
                    {
                        "type": "string",
                        "description": "An optional query parameter 'film-title'",
                        "name": "film-title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "An optional query parameter 'actor-name'",
                        "name": "actor-name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "An optional query parameter 'limit' that limits total number of returned films. By default 'limit' = 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "An optional query parameter 'offset' that indicates how many records should be skipped while listing films. By default 'offset' = 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_entrypoint_http.searchFilmsResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_vaberof_vk-internship-task_pkg_http_protocols_apiv1.Response": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_entrypoint_http.actorFilm": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_entrypoint_http.deleteActorResponseBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_app_entrypoint_http.film": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.filmActor"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "internal_app_entrypoint_http.filmActor": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                }
            }
        },
...
        "internal_app_entrypoint_http.searchFilmsResponseBody": {
            "type": "object",
            "properties": {
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.film"
                    }
                }
            }
        },
        "internal_app_entrypoint_http.updateActorRequestBody": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "sex": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        9
                    ]
                }
            }
        },
        "internal_app_entrypoint_http.updateActorResponseBody": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_entrypoint_http.actorFilm"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}